Cleaned up a lot more stuff!!

+++ Streamus/Backend/Domain/Managers/VideoManager.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Reflection;
-using NHibernate.Exceptions;
 using Streamus.Backend.Dao;
-using Streamus.Backend.Domain.DataInterfaces;
+using Streamus.Backend.Domain.Interfaces;
 using log4net;
 
 namespace Streamus.Backend.Domain.Managers
@@ -18,14 +17,23 @@ namespace Streamus.Backend.Domain.Managers
             VideoDao = videoDao;
         }
 
-        public void SaveVideo(Video video)
+        public void Save(Video video)
         {
             try
             {
                 NHibernateSessionManager.Instance.BeginTransaction();
-
                 video.ValidateAndThrow();
-                VideoDao.SaveOrUpdate(video);
+
+                //  TODO: Is this hitting the database? Surely it's not.
+                Video videoInSession = VideoDao.Get(video.Id);
+
+                if (videoInSession == null)
+                {
+                    VideoDao.Save(video);
+                }
+
+                //  TODO: Currently no mechanism in place to UPDATE a saved Video object.
+                //  I think this is OK because Video objects should only be created or deleted.
 
                 NHibernateSessionManager.Instance.CommitTransaction();
             }
@@ -36,7 +44,7 @@ namespace Streamus.Backend.Domain.Managers
             }
         }
 
-        public void SaveVideos(List<Video> videos)
+        public void Save(IEnumerable<Video> videos)
         {
             try
             {
@@ -58,14 +66,18 @@ namespace Streamus.Backend.Domain.Managers
         }
 
         //  Videos won't be deleted very often. No one user is able to do it -- so doesn't take a userId. Needs to be an admin.
-        public void DeleteVideoById(string id)
+        public void Delete(string id)
         {
             try
             {
                 NHibernateSessionManager.Instance.BeginTransaction();
 
-                Video video = VideoDao.GetById(id);
-                VideoDao.Delete(video);
+                Video video = VideoDao.Get(id);
+
+                if (video != null)
+                {
+                    VideoDao.Delete(video);
+                }
 
                 NHibernateSessionManager.Instance.CommitTransaction();
             }
@@ -76,11 +88,11 @@ namespace Streamus.Backend.Domain.Managers
             }
         }
 
-        public Video GetById(string id)
+        public Video Get(string id)
         {
             try
             {
-                Video video = VideoDao.GetById(id);
+                Video video = VideoDao.Get(id);
                 return video;
             }
             catch (Exception exception)
@@ -90,11 +102,11 @@ namespace Streamus.Backend.Domain.Managers
             }
         }
 
-        public IList<Video> GetByIds(List<string> ids)
+        public IList<Video> Get(List<string> ids)
         {
             try
             {
-                IList<Video> videos = VideoDao.GetByIds(ids);
+                IList<Video> videos = VideoDao.Get(ids);
                 return videos;
             }
             catch (Exception exception)
